[improve][txn] Take first snapshot before persisting the first transactional message (#21406)
### Motivation
The decision to write a snapshot before the first transaction message instead of before building the producer, is based on the fact that only the act of writing transactional messages signifies the use of the transaction buffer. Furthermore, it is only appropriate to schedule snapshot updates after this point.
Otherwise, it will add a lot of unnecessary IO read and write operations and increase the delay of topic load.

---
  **Scenario**
 * 1000 topics under namespace1.
 * Client1 enables transaction and sends transaction messages to topic 1.
 * Client1 sends normal messages to topic 2~500. 
 * Client2 disables transaction and sends messages to topic 501~1000. 
 
 **Internal Behavior**
 * Topic 1~500 will start the 500 task to write snapshots into the same system topic, e.g., system topic 1.
 * All the topics (1~1000) will read this system topic 1 when topic loading.
### Modifications

This Pull Request aims to resolve the unnecessary write operation. Starting to write snapshots when sending first transaction messages instead of building producer.